# event-booking-adapter-inbound-rest/src/main/java/com/musalasoft/eventbooking/rest/api/AuthApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.musalasoft.eventbooking.rest.api;

import com.musalasoft.eventbooking.rest.model.Credentials;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2024-08-10T15:36:30.869585+03:00[Africa/Cairo]")
@Validated
@Tag(name = "auth", description = "the auth API")
@RequestMapping("${openapi.eventBooking.base-path:}")
public interface AuthApi {

    /**
     * POST /auth : Authenticate a user.
     * This endpoint allows users to authenticate and receive a Bearer token.
     *
     * @param credentials The user&#39;s credentials. (optional)
     * @return OK (status code 200)
     *         or Unauthorized (status code 401)
     */
    @Operation(
        operationId = "authenticate",
        summary = "Authenticate a user.",
        tags = { "users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK"),
            @ApiResponse(responseCode = "401", description = "Unauthorized")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/auth",
        consumes = { "application/json" }
    )
    ResponseEntity<Void> authenticate(
        @Parameter(name = "credentials", description = "The user's credentials.") @Valid @RequestBody(required = false) Credentials credentials
    );

}