# event-booking-adapter-inbound-rest/src/main/java/com/musalasoft/eventbooking/rest/api/UsersApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.musalasoft.eventbooking.rest.api;

import com.musalasoft.eventbooking.rest.model.User;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2024-08-10T15:36:30.869585+03:00[Africa/Cairo]")
@Validated
@Tag(name = "users", description = "the users API")
@RequestMapping("${openapi.eventBooking.base-path:}")
public interface UsersApi {

    /**
     * POST /users : Create a new user.
     * This endpoint allows customers to create a new user.
     *
     * @param user The user to create. (optional)
     * @return Created (status code 201)
     *         or Bad request (status code 400)
     */
    @Operation(
        operationId = "createUser",
        summary = "Create a new user.",
        tags = { "users" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Created"),
            @ApiResponse(responseCode = "400", description = "Bad request")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/users",
        consumes = { "application/json" }
    )
    ResponseEntity<Void> createUser(
        @Parameter(name = "user", description = "The user to create.") @Valid @RequestBody(required = false) User user
    );

}